split out the resolver init
